Reorganize project structure.
