--- .qmake.conf
@@ -1,3 +1,0 @@
-load(qt_build_config)
-
-MODULE_VERSION = 1.0.0
