+++ mdkmediaservice.json
@@ -1,0 +1,4 @@
+{
+    "Keys": ["mdkservice"],
+    "Services": ["org.qt-project.qt.mediaplayer"]
+}
