+++ metadatareadercontrol.cpp
