+++ metadatareadercontrol.h
