--- qtmultimedia-plugins-mdk.pro
@@ -1,2 +1,0 @@
-TEMPLATE = subdirs
-SUBDIRS += src
