+++ renderercontrol.cpp
