--- src/mediaservice/plugin.json
@@ -1,4 +1,0 @@
-{
-    "Keys": ["mdkservice"],
-    "Services": ["org.qt-project.qt.mediaplayer"]
-}
